Rename neural machine translation to text2sparql (#2955)
* Rename neural machine translation to text2sparql


* Import fix
