+++ examples/nlp/text2sparql/evaluate_text2sparql.py
@@ -42,7 +42,7 @@ import os
 import pytorch_lightning as pl
 from omegaconf import DictConfig, OmegaConf
 
-from nemo.collections.nlp.models.neural_machine_translation import NeuralMachineTranslationModel
+from nemo.collections.nlp.models.text2sparql import Text2SparqlModel
 from nemo.core.config import hydra_runner
 from nemo.utils import logging
 
@@ -51,7 +51,7 @@ from nemo.utils import logging
 def main(cfg: DictConfig) -> None:
     logging.info(f"Config:\n {OmegaConf.to_yaml(cfg)}")
     trainer = pl.Trainer(gpus=cfg.trainer.gpus)
-    nmt_model = NeuralMachineTranslationModel.restore_from(restore_path=cfg.model.nemo_path)
+    nmt_model = Text2SparqlModel.restore_from(restore_path=cfg.model.nemo_path)
     nmt_model.setup_test_data(cfg.model.test_ds)
     results = trainer.test(nmt_model)
