+++ nemo/collections/nlp/data/__init__.py
@@ -30,10 +30,8 @@ from nemo.collections.nlp.data.machine_translation.machine_translation_dataset i
     TarredTranslationDataset,
     TranslationDataset,
 )
-from nemo.collections.nlp.data.neural_machine_translation.neural_machine_translation_dataset import (
-    NeuralMachineTranslationDataset,
-)
 from nemo.collections.nlp.data.question_answering_squad.qa_dataset import SquadDataset
+from nemo.collections.nlp.data.text2sparql.text2sparql_dataset import Text2SparqlDataset
 from nemo.collections.nlp.data.token_classification.token_classification_dataset import (
     BertTokenClassificationDataset,
     BertTokenClassificationInferDataset,
