+++ nemo/collections/nlp/models/__init__.py
@@ -25,10 +25,8 @@ from nemo.collections.nlp.models.intent_slot_classification import IntentSlotCla
 from nemo.collections.nlp.models.language_modeling.bert_lm_model import BERTLMModel
 from nemo.collections.nlp.models.language_modeling.transformer_lm_model import TransformerLMModel
 from nemo.collections.nlp.models.machine_translation import MTEncDecModel
-from nemo.collections.nlp.models.neural_machine_translation.neural_machine_translation_model import (
-    NeuralMachineTranslationModel,
-)
 from nemo.collections.nlp.models.question_answering.qa_model import QAModel
+from nemo.collections.nlp.models.text2sparql.text2sparql_model import Text2SparqlModel
 from nemo.collections.nlp.models.text_classification import TextClassificationModel
 from nemo.collections.nlp.models.token_classification import PunctuationCapitalizationModel, TokenClassificationModel
 from nemo.collections.nlp.models.zero_shot_intent_recognition import ZeroShotIntentModel
